+++ Simulator PIC16F84/Instruktionen/IORWF.cs
@@ -36,11 +36,11 @@ namespace Simulator_PIC16F84.Instruktionen
  
             if(d)
             {
-                W.Value.GetRegisterMap().RegisterList[f].Value = result;
+                W.Value.GetRegisterMap().RegisterList[f].Value = (sbyte)result;
             }
             else
             {
-                W.Value.Value = result;
+                W.Value.Value = (sbyte)result;
             }
         }
     }
