+++ Simulator PIC16F84/RegisterFileMap.cs
@@ -69,12 +69,12 @@ namespace Simulator_PIC16F84
 
         internal void SetZeroBit()
         {
-            RegisterList[3].Value = RegisterList[3].Value | 0x04;
+            RegisterList[3].Value = (sbyte) (RegisterList[3].Value | 0x04);
         }
 
         internal void ResetZeroBit()
         {
-            RegisterList[3].Value = RegisterList[3].Value & 0xFB;
+            RegisterList[3].Value = (sbyte) (RegisterList[3].Value & 0xFB);
         }
 
         public bool getZeroBit()
@@ -87,22 +87,22 @@ namespace Simulator_PIC16F84
 
         public void SetPowerDownBit()
         {
-            RegisterList[3].Value = RegisterList[3].Value | 0x08;
+            RegisterList[3].Value = (sbyte) (RegisterList[3].Value | 0x08);
         }
 
         public void ResetPowerDownBit()
         {
-            RegisterList[3].Value = RegisterList[3].Value & 0xF7;
+            RegisterList[3].Value = (sbyte) (RegisterList[3].Value & 0xF7);
         }
 
         public void SetTimeOutBit()
         {
-            RegisterList[3].Value = RegisterList[3].Value | 0x10;
+            RegisterList[3].Value = (sbyte) (RegisterList[3].Value | 0x10);
         }
 
         public void ResetTimeOutBit()
         {
-            RegisterList[3].Value = RegisterList[3].Value & 0xEF;
+            RegisterList[3].Value = (sbyte) (RegisterList[3].Value & 0xEF);
         }
